template: Possibility to have multiple parameters with same name
In sphinx it is allowed to have some parameters with the same name
multiple times, sql_attr_timestamp, sql_attr_float etc. Due to the
current way the template was generated that wasn't possible, because the
yaml would over-write the variable's value.

Giving some more 'intelligence' to the template, and make it checks the
type 'string' or 'list' allow now to do so. Default variable have been
modified to reflect those changes.